backend/pyproject.toml
[tool.poetry]
name = "sport_house_api"
version = "0.1.0"
description = ""
authors = ["maximkovalev <mkovalevhse@yandex.ru>"]

[tool.poetry.dependencies]
python = "3.10.1"
Django = "4.0.7"
djangorestframework = "^3.13.1"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"
psycopg2 = "^2.9.3"
argon2-cffi = "^21.3.0"
black = "^22.6.0"
django-filter = "^22.1"
django-redis = "^5.2.0"
gunicorn = "^20.1.0"
uvicorn = "^0.18.2"
factory-boy = "^3.2.1"
excel2json = "^1.0.1"
django-queryset-csv = "^1.1.0"
requests = "^2.28.1"
pandas = "^1.4.3"
scipy = "^1.9.0"
lime = "^0.2.0"
xgboost = "^1.6.1"
django-colorfield = "^0.7.2"
django-admin-generator = "^2.2.1"
locust = "^2.11.1"
celery = "^5.2.7"
django-celery-beat = "^2.3.0"
redis = "^4.3.4"
watchgod = "^0.8.2"
sentry-sdk = "^1.9.8"
flower = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
